+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -934,6 +934,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Changed the color space of EmissiveColorLDR property on all shader. Was linear but should have been sRGB. Auto upgrade script handle the conversion.
 - Preparation pass for RTSSShadows to be supported by render graph.
 - Add tooltips with the full name of the (graphics) compositor properties to properly show large names that otherwise are clipped by the UI (case 1263590)
+- Composition layers without any sub layers are now cleared to black to avoid confusion (case 1265061).
 
 ## [7.1.1] - 2019-09-05